+++ src/@locale/components/buttons/revoke.js
@@ -1,0 +1,15 @@
+const revoke =
+{
+    fr :
+    {
+        title   : 'Révoquer' ,
+        tooltip : 'Révoquer'
+    } ,
+    en :
+    {
+        title   : 'Revoke' ,
+        tooltip : 'Revoke'
+    }
+}
+
+export default revoke ;